Use patch set number for branch name if no topic
When downloading a Gerrit patch set, the local branch is created with
the naming scheme `review/AUTHOR/TOPIC`. But if no topic was set on the
patch set, this was resulting in a branch name `review/AUTHOR/nil`. To
avoid this, the patch set number is now used instead if no topic is
available, so that the branch will be named `review/AUTHOR/NUMBER` in
such cases. This behavior is also consistent with the one of the
`git-review` tool.